requirements.txt
python-dotenv
langchain
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv
langchain-community
langchain-text-splitters
tree-sitter
tree-sitter-languages
tqdm
pinecone
requests
langchain-pinecone
fastapi
uvicorn
GitPython
upstash-redis